refactor: Replace useEffect with useQuery
-refactor: Seperate engine details components
-refactor: Seperate new engines page into different components
